[ALPS07900403] PQ: fix coverity issue
avoid overflow ALPS07900403

MTK-Commit-Id: b0dd0e3082d0ec4eea32dc319475d40170cf795c

CR-Id: ALPS07900403
Feature: mml_pq

+++ drivers/gpu/drm/mediatek/mml/mtk-mml-fg-fw.c
@@ -545,7 +545,7 @@ static s32 calc_scaling_lut(s32 plane, struct mml_pq_fg_alg_data *fg_data)
 
 		if (delta_x == 0)
 			delta_x = 1;
-		delta = (s64)(delta_y * ((65536 + (delta_x >> 1)) / delta_x));
+		delta = (s64)(delta_y * (s64)((65536 + (delta_x >> 1)) / delta_x));
 
 		for (x = 0; x < delta_x; x++)
 			scaling_lut[value[point] + x] =
